Change order_id type from u64 to u32

+++ client/test/schema.ts
@@ -320,7 +320,7 @@ export const ConsumeEventsSchema: borsh.Schema = {
 
 export const CancelOrderSchema: borsh.Schema = {
     struct : {
-        order_id: "u64",
+        order_id: "u32",
         side: "u8"
     }
 }
